Added forgotten file
This is an old copy of the file found in a legacy repo. It might not be
flawless, but the one supposed to be here is somehow lost. By being forced
to use dropbox it's also lost on all distributed filesystems.

IOW this is better than nothing, even if really frustrating.

Change-Id: I94ea9e51a3eeeaab646b2a4b5c68371ec4237568